Update BubbleSort.h
int to double

+++ BubbleSort.h
@@ -6,6 +6,6 @@ using namespace::std;
 class BubbleSort
 {
 public:
-	void sort(int array[], int size);
+	void sort(double array[], int size);
 };
